LeetCode 12:python3的除法会有小数的问题

+++ Algorithms/12.py
@@ -10,11 +10,11 @@ class Solution(object):
         hundred_digits = ['', 'C', 'CC', 'CCC', 'CD', 'D', 'DC', 'DCC', 'DCCC', 'CM']
         ten_digits = ['', 'X', 'XX', 'XXX', 'XL', 'L', 'LX', 'LXX', 'LXXX', 'XC']
         digits = ['', 'I', 'II', 'III', 'IV', 'V', 'VI', 'VII', 'VIII', 'IX']
-        a = num / 1000
+        a = int(num / 1000)
         num = num % 1000
-        b = num / 100
+        b = int(num / 100)
         num = num % 100
-        c = num / 10
+        c = int(num / 10)
         d = num % 10
         roman = thousand_digits[a] + hundred_digits[b] + ten_digits[c] + digits[d]
         return roman
